agents/requirements.txt
kafka-python==1.4.6
pandas==0.24.2
pytz==2019.1
Faker==1.0.7
sqlalchemy==1.3.5
psycopg2-binary==2.8.3
